Added listview for each list(recycler)

+++ app/src/main/java/no/hiof/andersax/basket/model/ListItem.kt
@@ -1,5 +1,5 @@
 package no.hiof.andersax.basket.model
 
-class ListItem (val itemName : String,val quantity : Int, val isChecked : Boolean, val price : Int) {
+class ListItem(val itemName: String, val quantity: Long, val isChecked: Boolean, val price: Long) {
 
 }
